requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
browser-cookie3==0.12.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cryptography==3.3.1
cycler==0.10.0
distlib==0.3.1
Django==3.1.5
djangorestframework==3.12.2
filelock==3.0.12
future==0.18.2
html5lib==1.1
idna==2.10
jeepney==0.6.0
joblib==1.0.0
keyring==21.8.0
kiwisolver==1.3.1
lz4==3.1.1
matplotlib==3.3.3
nltk==3.5
numpy==1.19.5
pandas==1.2.1
pbkdf2==1.3
pefile==2019.4.18
Pillow==8.1.0
pyaes==1.6.1
pycparser==2.20
pycryptodome==3.9.9
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
PySimpleGUI==4.33.0
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.12
scikit-learn==0.24.1
scipy==1.6.0
SecretStorage==3.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.56.0
urllib3==1.26.2
virtualenv==20.4.0
webdriver-manager==3.2.2
webencodings==0.5.1
